+++ registry/storage/driver/azure/azure_test.go
@@ -2,7 +2,6 @@ package azure
 
 import (
 	"context"
-	"fmt"
 	"math/rand"
 	"os"
 	"strings"
@@ -10,7 +9,6 @@ import (
 
 	storagedriver "github.com/distribution/distribution/v3/registry/storage/driver"
 	"github.com/distribution/distribution/v3/registry/storage/driver/testsuites"
-	"github.com/stretchr/testify/suite"
 )
 
 const (
@@ -22,7 +20,7 @@ const (
 )
 
 var azureDriverConstructor func() (storagedriver.StorageDriver, error)
-var skipCheck func() string
+var skipCheck func(tb testing.TB)
 
 func init() {
 	var (
@@ -69,39 +67,23 @@ func init() {
 	}
 
 	// Skip Azure storage driver tests if environment variable parameters are not provided
-	skipCheck = func() string {
+	skipCheck = func(tb testing.TB) {
+		tb.Helper()
+
 		if len(missing) > 0 {
-			return fmt.Sprintf("Must set %s environment variables to run Azure tests", strings.Join(missing, ", "))
+			tb.Skipf("Must set %s environment variables to run Azure tests", strings.Join(missing, ", "))
 		}
-		return ""
 	}
 }
 
-func newDriverSuite() *testsuites.DriverSuite {
-	return testsuites.NewDriverSuite(azureDriverConstructor, skipCheck)
-}
-
 func TestAzureDriverSuite(t *testing.T) {
-	suite.Run(t, newDriverSuite())
+	skipCheck(t)
+	testsuites.Driver(t, azureDriverConstructor)
 }
 
 func BenchmarkAzureDriverSuite(b *testing.B) {
-	benchsuite := testsuites.NewDriverBenchmarkSuite(newDriverSuite())
-	benchsuite.Suite.SetupSuite()
-	b.Cleanup(benchsuite.Suite.TearDownSuite)
-
-	b.Run("PutGetEmptyFiles", benchsuite.BenchmarkPutGetEmptyFiles)
-	b.Run("PutGet1KBFiles", benchsuite.BenchmarkPutGet1KBFiles)
-	b.Run("PutGet1MBFiles", benchsuite.BenchmarkPutGet1MBFiles)
-	b.Run("PutGet1GBFiles", benchsuite.BenchmarkPutGet1GBFiles)
-	b.Run("StreamEmptyFiles", benchsuite.BenchmarkStreamEmptyFiles)
-	b.Run("Stream1KBFiles", benchsuite.BenchmarkStream1KBFiles)
-	b.Run("Stream1MBFiles", benchsuite.BenchmarkStream1MBFiles)
-	b.Run("Stream1GBFiles", benchsuite.BenchmarkStream1GBFiles)
-	b.Run("List5Files", benchsuite.BenchmarkList5Files)
-	b.Run("List50Files", benchsuite.BenchmarkList50Files)
-	b.Run("Delete5Files", benchsuite.BenchmarkDelete5Files)
-	b.Run("Delete50Files", benchsuite.BenchmarkDelete50Files)
+	skipCheck(b)
+	testsuites.BenchDriver(b, azureDriverConstructor)
 }
 
 var letterRunes = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
@@ -115,9 +97,7 @@ func randStringRunes(n int) string {
 }
 
 func TestCommitAfterMove(t *testing.T) {
-	if skipCheck() != "" {
-		t.Skip(skipCheck())
-	}
+	skipCheck(t)
 
 	driver, err := azureDriverConstructor()
 	if err != nil {
