+++ registry/storage/driver/filesystem/driver_test.go
@@ -6,45 +6,24 @@ import (
 
 	storagedriver "github.com/distribution/distribution/v3/registry/storage/driver"
 	"github.com/distribution/distribution/v3/registry/storage/driver/testsuites"
-	"github.com/stretchr/testify/suite"
 )
 
-func newDriverSuite(tb testing.TB) *testsuites.DriverSuite {
+func newDriverConstructor(tb testing.TB) testsuites.DriverConstructor {
 	root := tb.TempDir()
 
-	drvr, err := FromParameters(map[string]interface{}{
-		"rootdirectory": root,
-	})
-	if err != nil {
-		panic(err)
+	return func() (storagedriver.StorageDriver, error) {
+		return FromParameters(map[string]interface{}{
+			"rootdirectory": root,
+		})
 	}
-
-	return testsuites.NewDriverSuite(func() (storagedriver.StorageDriver, error) {
-		return drvr, nil
-	}, testsuites.NeverSkip)
 }
 
 func TestFilesystemDriverSuite(t *testing.T) {
-	suite.Run(t, newDriverSuite(t))
+	testsuites.Driver(t, newDriverConstructor(t))
 }
 
 func BenchmarkFilesystemDriverSuite(b *testing.B) {
-	benchsuite := testsuites.NewDriverBenchmarkSuite(newDriverSuite(b))
-	benchsuite.Suite.SetupSuite()
-	b.Cleanup(benchsuite.Suite.TearDownSuite)
-
-	b.Run("PutGetEmptyFiles", benchsuite.BenchmarkPutGetEmptyFiles)
-	b.Run("PutGet1KBFiles", benchsuite.BenchmarkPutGet1KBFiles)
-	b.Run("PutGet1MBFiles", benchsuite.BenchmarkPutGet1MBFiles)
-	b.Run("PutGet1GBFiles", benchsuite.BenchmarkPutGet1GBFiles)
-	b.Run("StreamEmptyFiles", benchsuite.BenchmarkStreamEmptyFiles)
-	b.Run("Stream1KBFiles", benchsuite.BenchmarkStream1KBFiles)
-	b.Run("Stream1MBFiles", benchsuite.BenchmarkStream1MBFiles)
-	b.Run("Stream1GBFiles", benchsuite.BenchmarkStream1GBFiles)
-	b.Run("List5Files", benchsuite.BenchmarkList5Files)
-	b.Run("List50Files", benchsuite.BenchmarkList50Files)
-	b.Run("Delete5Files", benchsuite.BenchmarkDelete5Files)
-	b.Run("Delete50Files", benchsuite.BenchmarkDelete50Files)
+	testsuites.BenchDriver(b, newDriverConstructor(b))
 }
 
 func TestFromParametersImpl(t *testing.T) {
