+++ registry/storage/driver/s3-aws/s3_test.go
@@ -16,8 +16,6 @@ import (
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/s3"
-	"github.com/stretchr/testify/suite"
-
 	"github.com/distribution/distribution/v3/internal/dcontext"
 	storagedriver "github.com/distribution/distribution/v3/registry/storage/driver"
 	"github.com/distribution/distribution/v3/registry/storage/driver/testsuites"
@@ -25,7 +23,7 @@ import (
 
 var (
 	s3DriverConstructor func(rootDirectory, storageClass string) (*Driver, error)
-	skipS3              func() string
+	skipCheck           func(tb testing.TB)
 )
 
 func init() {
@@ -142,49 +140,35 @@ func init() {
 	}
 
 	// Skip S3 storage driver tests if environment variable parameters are not provided
-	skipS3 = func() string {
+	skipCheck = func(tb testing.TB) {
+		tb.Helper()
+
 		if accessKey == "" || secretKey == "" || region == "" || bucket == "" || encrypt == "" {
-			return "Must set AWS_ACCESS_KEY, AWS_SECRET_KEY, AWS_REGION, S3_BUCKET, and S3_ENCRYPT to run S3 tests"
+			tb.Skip("Must set AWS_ACCESS_KEY, AWS_SECRET_KEY, AWS_REGION, S3_BUCKET, and S3_ENCRYPT to run S3 tests")
 		}
-		return ""
 	}
 }
 
-func newDriverSuite(tb testing.TB) *testsuites.DriverSuite {
+func newDriverConstructor(tb testing.TB) testsuites.DriverConstructor {
 	root := tb.TempDir()
 
-	return testsuites.NewDriverSuite(func() (storagedriver.StorageDriver, error) {
+	return func() (storagedriver.StorageDriver, error) {
 		return s3DriverConstructor(root, s3.StorageClassStandard)
-	}, skipS3)
+	}
 }
 
 func TestS3DriverSuite(t *testing.T) {
-	suite.Run(t, newDriverSuite(t))
+	skipCheck(t)
+	testsuites.Driver(t, newDriverConstructor(t))
 }
 
 func BenchmarkS3DriverSuite(b *testing.B) {
-	benchsuite := testsuites.NewDriverBenchmarkSuite(newDriverSuite(b))
-	benchsuite.Suite.SetupSuite()
-	b.Cleanup(benchsuite.Suite.TearDownSuite)
-
-	b.Run("PutGetEmptyFiles", benchsuite.BenchmarkPutGetEmptyFiles)
-	b.Run("PutGet1KBFiles", benchsuite.BenchmarkPutGet1KBFiles)
-	b.Run("PutGet1MBFiles", benchsuite.BenchmarkPutGet1MBFiles)
-	b.Run("PutGet1GBFiles", benchsuite.BenchmarkPutGet1GBFiles)
-	b.Run("StreamEmptyFiles", benchsuite.BenchmarkStreamEmptyFiles)
-	b.Run("Stream1KBFiles", benchsuite.BenchmarkStream1KBFiles)
-	b.Run("Stream1MBFiles", benchsuite.BenchmarkStream1MBFiles)
-	b.Run("Stream1GBFiles", benchsuite.BenchmarkStream1GBFiles)
-	b.Run("List5Files", benchsuite.BenchmarkList5Files)
-	b.Run("List50Files", benchsuite.BenchmarkList50Files)
-	b.Run("Delete5Files", benchsuite.BenchmarkDelete5Files)
-	b.Run("Delete50Files", benchsuite.BenchmarkDelete50Files)
+	skipCheck(b)
+	testsuites.BenchDriver(b, newDriverConstructor(b))
 }
 
 func TestEmptyRootList(t *testing.T) {
-	if skipS3() != "" {
-		t.Skip(skipS3())
-	}
+	skipCheck(t)
 
 	validRoot := t.TempDir()
 	rootedDriver, err := s3DriverConstructor(validRoot, s3.StorageClassStandard)
@@ -228,9 +212,7 @@ func TestEmptyRootList(t *testing.T) {
 }
 
 func TestStorageClass(t *testing.T) {
-	if skipS3() != "" {
-		t.Skip(skipS3())
-	}
+	skipCheck(t)
 
 	rootDir := t.TempDir()
 	contents := []byte("contents")
@@ -292,9 +274,7 @@ func TestStorageClass(t *testing.T) {
 }
 
 func TestDelete(t *testing.T) {
-	if skipS3() != "" {
-		t.Skip(skipS3())
-	}
+	skipCheck(t)
 
 	rootDir := t.TempDir()
 
@@ -493,9 +473,7 @@ func TestDelete(t *testing.T) {
 }
 
 func TestWalk(t *testing.T) {
-	if skipS3() != "" {
-		t.Skip(skipS3())
-	}
+	skipCheck(t)
 
 	rootDir := t.TempDir()
 
@@ -734,9 +712,7 @@ func TestWalk(t *testing.T) {
 }
 
 func TestOverThousandBlobs(t *testing.T) {
-	if skipS3() != "" {
-		t.Skip(skipS3())
-	}
+	skipCheck(t)
 
 	rootDir := t.TempDir()
 	standardDriver, err := s3DriverConstructor(rootDir, s3.StorageClassStandard)
@@ -762,9 +738,7 @@ func TestOverThousandBlobs(t *testing.T) {
 }
 
 func TestMoveWithMultipartCopy(t *testing.T) {
-	if skipS3() != "" {
-		t.Skip(skipS3())
-	}
+	skipCheck(t)
 
 	rootDir := t.TempDir()
 	d, err := s3DriverConstructor(rootDir, s3.StorageClassStandard)
@@ -815,9 +789,7 @@ func TestMoveWithMultipartCopy(t *testing.T) {
 }
 
 func TestListObjectsV2(t *testing.T) {
-	if skipS3() != "" {
-		t.Skip(skipS3())
-	}
+	skipCheck(t)
 
 	rootDir := t.TempDir()
 	d, err := s3DriverConstructor(rootDir, s3.StorageClassStandard)
